Bake in the library name.
